requirements.txt
streamlit==1.28.1
pandas==2.0.3
plotly==5.17.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
openpyxl==3.1.2
groq==0.4.2
